requirements.txt
google-cloud-speech>=2.21.0
